requirements.txt
# requirements.txt para empacotar o projeto com PyInstaller
streamlit
pandas
tabula-py
PyPDF2
openpyxl
